Gyroscope permission request added
