+++ index.css
@@ -57,8 +57,23 @@ body {
     font-size: 3.6vw;
     /* font-size: 20px; */
     background-color: #7a5f41c7;
+    margin-right: 3.6vw;  /* Add a 20 pixel gap to the right */
+}
+#gyro-toggle{
+    height: 9.9vw;
+    width: 24.3vw;
+    border-radius: 1.44vw;
+    border-style: none;
+    outline: none;
+
+    color: white;
+    font-weight: bold;
+    font-size: 3.6vw;
+    background-color: #7a5f41c7;
+    margin-right: 3.6vw;  /* Add a 20 pixel gap to the right */
     
 }
+
 
 #score_container {
     display: flex;
@@ -213,6 +228,21 @@ body {
         font-weight: bold;
         font-size: 10px;
         background-color: #7a5f41c7;
+        margin-right: 10px;  /* Add a 20 pixel gap to the right */
+        
+    }
+    #gyro-toggle{
+        height: 27.5px;
+        width: 67.5px;
+        border-radius: 4px;
+        border-style: none;
+        outline: none;
+    
+        color: white;
+        font-weight: bold;
+        font-size: 10px;
+        background-color: #7a5f41c7;
+        margin-right: 10px;  /* Add a 20 pixel gap to the right */
         
     }
     
@@ -353,6 +383,21 @@ body {
         font-weight: bold;
         font-size: 12px;
         background-color: #7a5f41c7;
+        margin-right: 12px;  /* Add a 20 pixel gap to the right */
+        
+    }
+    #gyro-toggle{
+        height: 33px;
+        width: 81px;
+        border-radius: 4.8px;
+        border-style: none;
+        outline: none;
+    
+        color: white;
+        font-weight: bold;
+        font-size: 12px;
+        background-color: #7a5f41c7;
+        margin-right: 12px;  /* Add a 20 pixel gap to the right */
         
     }
     
@@ -491,6 +536,22 @@ body {
         font-weight: bold;
         font-size: 20px;
         background-color: #7a5f41c7;
+        margin-right: 20px;  /* Add a 20 pixel gap to the right */
+        
+    }
+
+    #gyro-toggle{
+        height: 55px;
+        width: 135px;
+        border-radius: 8px;
+        border-style: none;
+        outline: none;
+    
+        color: white;
+        font-weight: bold;
+        font-size: 20px;
+        background-color: #7a5f41c7;
+        margin-right: 20px;  /* Add a 20 pixel gap to the right */
         
     }
