+++ index.html
@@ -23,6 +23,9 @@
         <button id="btn_content">
             RESTART
         </button>
+        <button id="gyro-toggle">
+            Gyroscope: <span id="gyro-status">Disabled</span>
+        </button>
     </div>
 
     <!-- 分数显示 -->
