+++ index.js
@@ -6,8 +6,8 @@ import handleKeydown from './happy.NIU/handelKeydown.js';
 //Yable: import handleSwipeMove 
 import handleSwipeMove from './happy.NIU/handleSwipeMove.js'
 //Yable: import GyroscopeControl
-import handleOrientation from './happy.NIU/handleOrientation.js'
-
+import handleOrientation from './happy.NIU/GyroscopeControl.js'
+import requestOrientationPermission from './happy.NIU/GyroscopeControl.js'
 // 创建一个4*4的二维数组作为棋盘
 // 里面的每一个值都是二的n次幂，其中n>=1，而且可以等于零
 export const gameBox = []
@@ -22,6 +22,10 @@ export const shareGameOver = {
 export const restart = document.getElementById('btn_content');
 export const board = document.getElementById('square_container');
 
+//gyroscope status
+var isGyroscopeEnabled = false;
+var gyroToggle = document.getElementById("gyro-toggle");
+var gyroStatus = document.getElementById("gyro-status");
 
 // 创建方块对象
 // 对象名：square
@@ -51,13 +55,23 @@ export function gameStart() {
     document.addEventListener('touchstart', handleSwipeMove, false);
     document.addEventListener('touchmove', handleSwipeMove, false);
     document.addEventListener('touchend', handleSwipeMove, false);
+    
     //Yable: add a gyroscope listener
-    if (window.DeviceOrientationEvent) {
-        window.addEventListener('deviceorientation', handleOrientation, false);
-    } else {
-        // Device Orientation isn't supported
-        console.log("Device Orientation is not supported.")
-    }
+    gyroToggle.addEventListener("click", function() {
+        isGyroscopeEnabled = !isGyroscopeEnabled;
+        gyroStatus.textContent = isGyroscopeEnabled ? "Enabled" : "Disabled";
+    
+        if (isGyroscopeEnabled) {
+            // Add the deviceorientation event listener
+            requestOrientationPermission();
+        } else {
+            // Remove the deviceorientation event listener
+            window.removeEventListener('deviceorientation', handleOrientation, false);
+            }
+        })
+                
+
+    
  
 
     // 如果按了重新开始按钮就进行初始化
